TXM-10717 Update R2-SB branch to use Aristocrat.Protocols 1.1.9 version Merge in main as of 03/14/2023

+++ Protocol/Bingo/Aristocrat.Bingo.Client/Messages/ReportEventMessage.cs
@@ -2,13 +2,23 @@
 {
     using System;
 
+    /// <summary>
+    ///     The message for reporting events to the server
+    /// </summary>
     [Serializable]
     public class ReportEventMessage : IMessage
     {
+        /// <summary>
+        ///     Creates an instance of <see cref="ReportEventMessage"/>
+        /// </summary>
+        /// <param name="machineSerial">The machine serial for this event</param>
+        /// <param name="timeStamp">The timestamp for this event</param>
+        /// <param name="eventId">The event ID for this event</param>
+        /// <param name="eventType">The event type for this event</param>
         public ReportEventMessage(
             string machineSerial,
             DateTime timeStamp,
-            int eventId,
+            long eventId,
             int eventType)
         {
             MachineSerial = machineSerial;
@@ -17,12 +27,24 @@
             EventType = eventType;
         }
 
+        /// <summary>
+        ///     Gets the machine serial
+        /// </summary>
         public string MachineSerial { get; set; }
 
+        /// <summary>
+        ///     Gets the timestamp for this event
+        /// </summary>
         public DateTime TimeStamp { get; }
 
-        public int EventId { get; }
+        /// <summary>
+        ///     Gets the event ID for this event
+        /// </summary>
+        public long EventId { get; }
 
+        /// <summary>
+        ///     Gets the event type for this event
+        /// </summary>
         public int EventType { get; }
     }
 }
